Changes: - allow domain names as well as IPs - Skip files if already downloaded (saves bandwith and time) - Can delete files on the SD card after downloaded (avoids filling the SD card until no space left)

+++ VstarCamDownloader.py
@@ -1,10 +1,12 @@
-#DownloadVideos v1.0.0 by RB
-#DownloadVideos.py 192.168.0.15 80 /opt/videos/
+#VstarCamDownloader v1.0.0 by RB
+#Edited by Sergio Faustino
+#Usage Example: VstarCamDownloader.py cameraip.changeip.net 28080 videos/IPCam1 MOVE admin 88888888
 
 import sys
 import os
 import urllib2
 import base64
+import socket
 
 def downloadfile(url, filename, directory, username, password):
 
@@ -12,40 +14,61 @@ def downloadfile(url, filename, directory, username, password):
     base64string = base64.encodestring('%s:%s' % (username, password)).replace('\n', '')
     request.add_header("Authorization", "Basic %s" % base64string)   
     up = urllib2.urlopen(request)
-    fp = open(directory + '/' + filename, 'wb');
     meta = up.info();
-
     file_size = int(meta.getheaders("Content-Length")[0])
-    print "Downloading: %s Bytes: %s" % (filename, file_size)
 
-    file_size_dl = 0
-    block_sz = 8192
-    while True:
-        buffer = up.read(block_sz)
-        if not buffer:
-            break
+    #Check if file already downloaded
+    if os.path.exists(directory + '/' + filename) and os.path.getsize(directory + '/' + filename) == file_size:
+        print 'Already downloaded, skipping file ' + filename + ' with size ' + str(file_size);
+    else:
+        fp = open(directory + '/' + filename, 'wb');
+        print "Downloading: %s Bytes: %s" % (filename, file_size)
 
-        file_size_dl += len(buffer)
-        fp.write(buffer)
-        status = r"%10d  [%3.2f%%]" % (file_size_dl, file_size_dl * 100. / file_size)
-        status = status + chr(8)*(len(status)+1)
-        print status,
-    fp.close()
+        file_size_dl = 0
+        block_sz = 8192
+        while True:
+            buffer = up.read(block_sz)
+            if not buffer:
+                break
 
+            file_size_dl += len(buffer)
+            fp.write(buffer)
+            status = r"%10d  [%3.2f%%]" % (file_size_dl, file_size_dl * 100. / file_size)
+            status = status + chr(8)*(len(status)+1)
+            print status,
+        fp.close()
+
+def deletefile(fileurl):
+    try:
+        response = urllib2.urlopen(fileurl, timeout = 10);
+    except urllib2.URLError as e:
+        print "Error - Failed to access URL";
+        sys.exit(5);
+    except socket.timeout as e:
+        print "Error - Failed to access URL";
+        sys.exit(6);
+    except:
+        print "Error - Is this IP a vstarcam camera?";
+        sys.exit(7)
+
+    print "Deleted file " + file
+ 
 def isgoodipv4(s):
-    pieces = s.split('.')
+    ip = socket.gethostbyname(s)
+    pieces = ip.split('.')
     if len(pieces) != 4: return False
     try: return all(0<=int(p)<256 for p in pieces)
     except ValueError: return False
 
 if len(sys.argv) < 6:
-  print "DownloadVideos.py [IP] [PORT] [OUTPUT DIRECTORY] [USERNAME] [PASSWORD]";
+  print "VstarCamDownloader.py [IP] [PORT] [OUTPUT DIRECTORY] [COPY/MOVE] [USERNAME] [PASSWORD]";
   sys.exit(1);
 
-ip   = sys.argv[1];
+ip = sys.argv[1];
 directory = sys.argv[3];
-username = sys.argv[4];
-password = sys.argv[5];
+operation = sys.argv[4];
+username = sys.argv[5];
+password = sys.argv[6];
 
 try:
     port = int(sys.argv[2]);
@@ -104,8 +127,13 @@ for line in lines:
         filelength.append(line[-10:-2]);
 
 #Get files and put them into the output directory
+print 'Found ' + str(len(files)) + ' files';
 for file in files:
     fileurl = 'http://' + ip + ':' + str(port) + '/record/' + file;
     downloadfile(fileurl, file, directory, username, password);
 
-print "Downloaded files successfully to " + directory;
+    if operation == 'MOVE':
+        fileurl = 'http://' + ip + ':' + str(port) + '/del_file.cgi?name=' + file + '&loginuse=' + username + '&loginpas=' + password;
+        deletefile(fileurl);
+
+print "Finished downloading files to " + directory;
